Start refactor from page -> component
