+++ src/components/terminate-test-modal/terminate-test-modal.scss
